Implemented file caching and rerouting
As if now the rerouting will only happen to index.htm.
In the future mroe default landingpages will be searched, and if the
page is hot, a gziped version will be created and used.
Warning, the modified field is not yet used to check if the cache is
older than the current file, this will be implemented later